+++ doc/ReleaseNotes.md
@@ -7,6 +7,7 @@ These release notes start with the October 18, 2021 CloudComPy release.
  - Update to CloudCompare 2.12.0 (april 7 2022)
  - Add access to plugin RANSAC_SD on Windows and Linux (issue #27). See test035.py.
  - Cross-section methods (issue #27). See test036.py
+ - Extract Connected Components (issue #27). See test037.py
  - fix tests, activate trace on all tests to improve debug
  - add methods on primitives
  - fix a bug in conversion from C++ to Python, where objects name are missing.